feature: android navigation controller
