--- ui/navigationcontroller/index.js
@@ -1,5 +1,0 @@
-// if (Device.deviceOS === "iOS") {
-//     module.exports = require('./navigationcontroller-iOS');
-// } else if (Device.deviceOS === "Android") {
-module.exports = require("./navigationcontroller-" + Device.deviceOS);
-// }
